Use accompanist-flowlayout for ChipRows again
