+++ ui/detail/src/main/java/de/schnettler/scrobbler/ui/detail/screen/ArtistDetails.kt
@@ -12,6 +12,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.google.accompanist.insets.navigationBarsHeight
 import de.schnettler.common.whenNotEmpty
 import de.schnettler.database.models.EntityWithStats.TrackWithStats
 import de.schnettler.database.models.EntityWithStatsAndInfo.ArtistWithStatsAndInfo
@@ -22,6 +23,7 @@ import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction
 import de.schnettler.scrobbler.ui.common.compose.navigation.UIAction.ListingSelected
 import de.schnettler.scrobbler.ui.common.compose.rememberDominantColorCache
 import de.schnettler.scrobbler.ui.common.compose.widget.Carousel
+import de.schnettler.scrobbler.ui.common.compose.widget.ChipFlowRow
 import de.schnettler.scrobbler.ui.common.compose.widget.CollapsingToolbar
 import de.schnettler.scrobbler.ui.common.compose.widget.Header
 import de.schnettler.scrobbler.ui.common.compose.widget.ListeningStats
@@ -29,9 +31,7 @@ import de.schnettler.scrobbler.ui.common.compose.widget.MediaCard
 import de.schnettler.scrobbler.ui.common.compose.widget.PlainListIconBackground
 import de.schnettler.scrobbler.ui.common.util.abbreviate
 import de.schnettler.scrobbler.ui.detail.R
-import de.schnettler.scrobbler.ui.detail.widget.ChipRow
 import de.schnettler.scrobbler.ui.detail.widget.ExpandingInfoCard
-import com.google.accompanist.insets.navigationBarsHeight
 
 @Composable
 fun ArtistDetailScreen(
@@ -69,7 +69,7 @@ private fun LazyListScope.detailItems(
     info?.tags?.whenNotEmpty { tags ->
         itemSpacer(16.dp)
         item { Header(title = stringResource(id = R.string.header_tags)) }
-        item { ChipRow(items = tags, onChipClicked = { actioner(UIAction.TagSelected(it)) }) }
+        item { ChipFlowRow(items = tags, onChipClicked = { actioner(UIAction.TagSelected(it)) }) }
     }
 
     // Tracks
