+++ ui/detail/src/main/java/de/schnettler/scrobbler/ui/detail/screen/TrackDetails.kt
@@ -24,7 +24,7 @@ import de.schnettler.scrobbler.ui.common.compose.widget.ListeningStats
 import de.schnettler.scrobbler.ui.detail.AlbumCategory
 import de.schnettler.scrobbler.ui.detail.R
 import de.schnettler.scrobbler.ui.detail.viewmodel.TrackViewModel
-import de.schnettler.scrobbler.ui.detail.widget.ChipRow
+import de.schnettler.scrobbler.ui.common.compose.widget.ChipRow
 import de.schnettler.scrobbler.ui.detail.widget.ExpandingInfoCard
 import com.google.accompanist.insets.navigationBarsHeight
 import com.google.accompanist.insets.navigationBarsPadding
